+++ R/utils-persons.R
@@ -1,9 +1,6 @@
 bibtex_pers_von_last_first_jr <- function(x) {
   # Protect commas on brackets to avoid error on splitting
-  protected <- gsub(",(?![^\\}]*(\\{|$))", "@comma@",
-    x,
-    perl = TRUE
-  )
+  protected <- gsub(",(?![^\\}]*(\\{|$))", "@comma@", x, perl = TRUE)
 
   parts_comma <- trimws(unlist(strsplit(protected, ",")))
 
@@ -20,13 +17,14 @@ bibtex_pers_von_last_first_jr <- function(x) {
   # Junior part
   jr <- parts_comma[2]
 
-
   # Now it is the same than in bibtex_person_von_last_first
 
   # Now identify the von part
 
   # Protect spaces between braces before splitting
-  protected <- gsub("\\s(?![^\\}]*(\\{|$))", "@blank@",
+  protected <- gsub(
+    "\\s(?![^\\}]*(\\{|$))",
+    "@blank@",
     parts_comma[1],
     perl = TRUE
   )
@@ -39,17 +37,14 @@ bibtex_pers_von_last_first_jr <- function(x) {
   # Start building
   # Assess casing
 
-  is_upper <- vapply(parts,
-    FUN.VALUE = logical(1),
-    function(y) {
-      case <- substr(y, 1, 1)
-      if (case == toupper(case)) {
-        TRUE
-      } else {
-        FALSE
-      }
+  is_upper <- vapply(parts, FUN.VALUE = logical(1), function(y) {
+    case <- substr(y, 1, 1)
+    if (case == toupper(case)) {
+      TRUE
+    } else {
+      FALSE
     }
-  )
+  })
 
   # Family should be always provided
   family <- parts[length(parts)]
@@ -59,15 +54,19 @@ bibtex_pers_von_last_first_jr <- function(x) {
   # Get more family parts
   invert <- rev(upper)
   # Then need to get sequentally the family
-  for (i in seq_len(length(invert))) {
-    if (invert[i] == FALSE) break
+  for (i in seq_along(invert)) {
+    if (invert[i] == FALSE) {
+      break
+    }
     family <- c(names(invert[i]), family)
   }
 
   # The remaining is von
   von <- setdiff(names(is_upper), family)
 
-  if (length(von) == 0) von <- NULL
+  if (length(von) == 0) {
+    von <- NULL
+  }
 
   # Compose final list to be pased to person()
 
@@ -79,21 +78,17 @@ bibtex_pers_von_last_first_jr <- function(x) {
     jr = jr
   )
 
-  return(end_list)
+  end_list
 }
 
 bibtex_pers_von_last_first <- function(x) {
   # Protect commas on brackets to avoid error on splitting
-  protected <- gsub(",(?![^\\}]*(\\{|$))", "@comma@",
-    x,
-    perl = TRUE
-  )
+  protected <- gsub(",(?![^\\}]*(\\{|$))", "@comma@", x, perl = TRUE)
 
   parts_comma <- trimws(unlist(strsplit(protected, ",")))
 
   # Unprotect
   parts_comma <- gsub("@comma@", ",", parts_comma)
-
 
   if (length(parts_comma) == 2) {
     given <- parts_comma[2]
@@ -102,11 +97,12 @@ bibtex_pers_von_last_first <- function(x) {
     given <- NULL
   }
 
-
   # Now identify the von part
 
   # Protect spaces between braces before splitting
-  protected <- gsub("\\s(?![^\\}]*(\\{|$))", "@blank@",
+  protected <- gsub(
+    "\\s(?![^\\}]*(\\{|$))",
+    "@blank@",
     parts_comma[1],
     perl = TRUE
   )
@@ -119,17 +115,14 @@ bibtex_pers_von_last_first <- function(x) {
   # Start building
   # Assess casing
 
-  is_upper <- vapply(parts,
-    FUN.VALUE = logical(1),
-    function(y) {
-      case <- substr(y, 1, 1)
-      if (case == toupper(case)) {
-        TRUE
-      } else {
-        FALSE
-      }
+  is_upper <- vapply(parts, FUN.VALUE = logical(1), function(y) {
+    case <- substr(y, 1, 1)
+    if (case == toupper(case)) {
+      TRUE
+    } else {
+      FALSE
     }
-  )
+  })
 
   # Family should be always provided
   family <- parts[length(parts)]
@@ -139,15 +132,19 @@ bibtex_pers_von_last_first <- function(x) {
   # Get more family parts
   invert <- rev(upper)
   # Then need to get sequentally the family
-  for (i in seq_len(length(invert))) {
-    if (invert[i] == FALSE) break
+  for (i in seq_along(invert)) {
+    if (invert[i] == FALSE) {
+      break
+    }
     family <- c(names(invert[i]), family)
   }
 
   # The remaining is von
   von <- setdiff(names(is_upper), family)
 
-  if (length(von) == 0) von <- NULL
+  if (length(von) == 0) {
+    von <- NULL
+  }
 
   # Compose final list to be pased to person()
 
@@ -158,10 +155,8 @@ bibtex_pers_von_last_first <- function(x) {
     family = family
   )
 
-  return(end_list)
+  end_list
 }
-
-
 
 
 bibtex_pers_first_von_last <- function(x) {
@@ -179,58 +174,48 @@ bibtex_pers_first_von_last <- function(x) {
   # Jean de La Fontaine     -> "Jean"         "de"          "La Fontaine"
 
   # Protect spaces between braces before splitting
-  x <- gsub("\\s(?![^\\}]*(\\{|$))", "@blank@",
-    x,
-    perl = TRUE
-  )
+  x <- gsub("\\s(?![^\\}]*(\\{|$))", "@blank@", x, perl = TRUE)
 
   # Collapse blanks
   x <- gsub("\\s+", " ", x)
 
   parts <- trimws(unlist(strsplit(x, " ")))
 
-
   # Unprotect
   parts <- gsub("@blank@", " ", parts)
-
-
 
   # Start building
   # Assess casing
 
-  is_upper <- vapply(parts,
-    FUN.VALUE = logical(1),
-    function(y) {
-      case <- substr(y, 1, 1)
-      if (case == toupper(case)) {
-        TRUE
-      } else {
-        FALSE
-      }
+  is_upper <- vapply(parts, FUN.VALUE = logical(1), function(y) {
+    case <- substr(y, 1, 1)
+    if (case == toupper(case)) {
+      TRUE
+    } else {
+      FALSE
     }
-  )
+  })
 
   # Family should be always provided
   family <- parts[length(parts)]
-
 
   # Get if there is a mix of casings on the remaining parts
   upper <- is_upper[setdiff(names(is_upper), family)]
   mix <- (length(unique(upper)) == 1)
 
-
   if (!all(mix, upper)) {
     invert <- rev(upper)
     # Then need to get sequentally the family
-    for (i in seq_len(length(invert))) {
-      if (invert[i] == FALSE) break
+    for (i in seq_along(invert)) {
+      if (invert[i] == FALSE) {
+        break
+      }
       family <- c(names(invert[i]), family)
     }
   }
 
   # Clean again
   upper <- is_upper[setdiff(names(is_upper), family)]
-
 
   # von part
   # Detect casing index on remaining parts
@@ -244,7 +229,6 @@ bibtex_pers_first_von_last <- function(x) {
     von <- names(upper[seq(min(von_assess), max(von_assess))])
   }
 
-
   # get remaining part, it should be First
   given <- names(upper[setdiff(names(upper), von)])
 
@@ -257,7 +241,7 @@ bibtex_pers_first_von_last <- function(x) {
     family = family
   )
 
-  return(end_list)
+  end_list
 }
 
 validate_cff_person_fields <- function(person_cff) {
@@ -290,42 +274,24 @@ split_txt_persons <- function(person) {
 
   # Protect 'and' on brackets {}
   # Lower
-  protected <- gsub("(and)(?![^\\}]*(\\{|$))", "@nd@",
-    person,
-    perl = TRUE
-  )
+  protected <- gsub("(and)(?![^\\}]*(\\{|$))", "@nd@", person, perl = TRUE)
 
   # upper
-  protected <- gsub("AND(?![^\\}]*(\\{|$))", "@ND@",
-    protected,
-    perl = TRUE
-  )
+  protected <- gsub("AND(?![^\\}]*(\\{|$))", "@ND@", protected, perl = TRUE)
 
   # Do the same for 'and' in comments "()" as provided by print.person
   # Lower
-  protected <- gsub("(and)(?![^\\)]*(\\(|$))", "@nd@",
-    protected,
-    perl = TRUE
-  )
+  protected <- gsub("(and)(?![^\\)]*(\\(|$))", "@nd@", protected, perl = TRUE)
 
   # upper
-  protected <- gsub("AND(?![^\\)]*(\\(|$))", "@ND@",
-    protected,
-    perl = TRUE
-  )
+  protected <- gsub("AND(?![^\\)]*(\\(|$))", "@ND@", protected, perl = TRUE)
 
   # Do the same for 'and' in "<>". These are email, should never happen
   # Lower
-  protected <- gsub("(and)(?![^>]*(<|$))", "@nd@",
-    protected,
-    perl = TRUE
-  )
+  protected <- gsub("(and)(?![^>]*(<|$))", "@nd@", protected, perl = TRUE)
 
   # upper
-  protected <- gsub("AND(?![^>]*(<|$))", "@ND@",
-    protected,
-    perl = TRUE
-  )
+  protected <- gsub("AND(?![^>]*(<|$))", "@ND@", protected, perl = TRUE)
 
   auths <- unlist(strsplit(protected, " and | AND "))
 
@@ -337,6 +303,8 @@ split_txt_persons <- function(person) {
 }
 
 extract_person_comments <- function(person) {
+  # nocov start
+  # Not used in R >= 4.5.0
   # Ensure person type
   person <- as.person(person)
 
@@ -347,10 +315,11 @@ extract_person_comments <- function(person) {
   comment_as_text <- tolower(clean_str(comm_cff))
 
   # Special case when coerced from text, only can extract orcid and web
-  if (all(
-    any(is.na(nms_com), length(nms_com) == 0),
-    length(comment_as_text > 0)
-  )
+  if (
+    all(
+      any(is.na(nms_com), length(nms_com) == 0),
+      length(comment_as_text > 0)
+    )
   ) {
     split_comments <- unlist(strsplit(comment_as_text, ",| |<|>"))
 
@@ -406,5 +375,82 @@ extract_person_comments <- function(person) {
   )
   fin_list$email <- clean_str(email)
 
+  fin_list
+  # nocov end
+}
+
+extract_person_comments45 <- function(person) {
+  # Detect comments
+  # Last is )?
+
+  if ("person" %in% class(person)) {
+    person <- format(person)
+  }
+  last <- substr(person, nchar(person), nchar(person))
+  # Has another )?
+  start_comment <- min(regexpr("(", person, fixed = TRUE))
+
+  has_comment <- all(last == ")", start_comment > 5)
+
+  # If does not have comment convert to person
+  if (!has_comment) {
+    p <- as.person(person)
+    comm_cff <- list(fake = 123)
+  } else {
+    # Split and convert
+    the_person <- substr(person, 1, start_comment - 1)
+    p <- as.person(the_person)
+
+    # Comments
+    # Can't get commas in the comment :(
+    comm <- strsplit(substr(person, start_comment + 1, nchar(person) - 1), ",")
+    comm <- unlist(comm)
+    comm_list <- lapply(comm, function(x) {
+      detect_split <- min(regexpr(":", x, fixed = TRUE))
+      nm <- substr(x, 1, detect_split - 1)
+      val <- substr(x, detect_split + 1, nchar(x))
+      val <- clean_str(val)
+      val <- gsub("<|>", "", val)
+
+      nm <- clean_str(nm)
+      names(val) <- nm
+      val
+    })
+
+    comm_cff <- as.list(unlist(comm_list))
+  }
+
+  names(comm_cff) <- tolower(names(comm_cff))
+  # Delete non-named comments
+  nms_com <- names(comm_cff)
+  comm_cff <- comm_cff[nchar(nms_com) > 1]
+
+  if (!is.null(comm_cff$orcid)) {
+    orcid <- gsub("^orcid.org/", "", comm_cff$orcid)
+    orcid <- gsub("^https://orcid.org/", "", orcid)
+    orcid <- gsub("^http://orcid.org/", "", orcid)
+
+    comm_cff$orcid <- paste0("https://orcid.org/", orcid)
+  }
+
+  # Add website
+  web <- comm_cff$website
+
+  if (!is.null(web)) {
+    comm_cff$website <- clean_str(web[is_url(web)])
+  }
+
+  # Add also email
+  # Check if several mails (MomTrunc 6.0)
+  look_emails <- c(unlist(p$email), comm_cff$email)
+  valid_emails <- unlist(lapply(look_emails, is_email))
+  email <- look_emails[valid_emails][1]
+
+  # Final list
+  fin_list <- c(
+    list(email = NULL),
+    comm_cff["email" != names(comm_cff)]
+  )
+  fin_list$email <- clean_str(email)
   fin_list
 }