+++ tests/testthat/test-cff_write_misc.R
@@ -3,10 +3,7 @@ test_that("Errors", {
 })
 
 test_that("Write", {
-  bib <- bibentry("Misc",
-    title = "My title",
-    author = "Fran Pérez"
-  )
+  bib <- bibentry("Misc", title = "My title", author = "Fran Pérez")
 
   file <- file.path(tempdir(), "noext")
   expect_message(cff_write_bib(bib, file, verbose = TRUE))
@@ -30,10 +27,7 @@ test_that("Write", {
 
 
 test_that("Write ASCII", {
-  bib <- bibentry("Misc",
-    title = "My title",
-    author = "Fran Pérez"
-  )
+  bib <- bibentry("Misc", title = "My title", author = "Fran Pérez")
 
   file <- file.path(tempdir(), "ascii.bib")
   expect_silent(cff_write_bib(bib, file, verbose = FALSE, ascii = TRUE))
@@ -44,10 +38,7 @@ test_that("Write ASCII", {
 })
 
 test_that("Test append", {
-  bib <- bibentry("Misc",
-    title = "My title",
-    author = "Fran Herrero"
-  )
+  bib <- bibentry("Misc", title = "My title", author = "Fran Herrero")
   cf <- system.file("examples/DESCRIPTION_basic", package = "cffr")
 
   a_cff <- cff_read(cf)
@@ -59,11 +50,12 @@ test_that("Test append", {
   lines1 <- readLines(file)
 
   # Append
-  bib2 <- bibentry("Misc",
-    key = "key2", title = "Another title",
+  bib2 <- bibentry(
+    "Misc",
+    key = "key2",
+    title = "Another title",
     author = "Ian Henderson"
   )
-
 
   cff_write_bib(bib2, file, verbose = FALSE, append = TRUE)
   cff_write_bib(a_cff, file, verbose = FALSE, append = TRUE)
@@ -72,7 +64,7 @@ test_that("Test append", {
   lines2 <- readLines(file)
 
   # First lines identical, more lines on append
-  expect_true(all(lines1 == lines2[seq_len(length(lines1))]))
+  expect_true(all(lines1 == lines2[seq_along(lines1)]))
 
   expect_gt(length(lines2), length(lines1))
 
@@ -80,16 +72,13 @@ test_that("Test append", {
   cff_write_bib(bib2, file, verbose = FALSE, append = FALSE)
   lines3 <- readLines(file)
 
-  expect_false(all(lines1 == lines3[seq_len(length(lines1))]))
+  expect_false(all(lines1 == lines3[seq_along(lines1)]))
   expect_lt(length(lines3), length(lines2))
 })
 
 
 test_that("Test dir creation", {
-  bib <- bibentry("Misc",
-    title = "My title",
-    author = "Fran Herrero"
-  )
+  bib <- bibentry("Misc", title = "My title", author = "Fran Herrero")
 
   file <- file.path(tempdir(), "idontexist", "append.bib")
 
@@ -124,10 +113,7 @@ test_that("Write CITATION", {
   thepath <- list.files(f, pattern = "DESCRIPTION_basicdate", full.names = TRUE)
   f1 <- cff_read(thepath)
 
-  bib <- bibentry("Misc",
-    title = "My title",
-    author = "Fran Pérez"
-  )
+  bib <- bibentry("Misc", title = "My title", author = "Fran Pérez")
 
   file <- file.path(tempdir(), "CITAT_ION")
   expect_message(cff_write_citation(bib, file, verbose = TRUE))
@@ -136,9 +122,12 @@ test_that("Write CITATION", {
   expect_false(file_exist_abort(paste0(file, ".bk1")))
 
   # Check now backup exists and use cff
-  expect_silent(cff_write_citation(f1, file,
+  expect_silent(cff_write_citation(
+    f1,
+    file,
     verbose = FALSE,
-    what = "all", append = TRUE
+    what = "all",
+    append = TRUE
   ))
 
   expect_snapshot_file(file)