+++ tests/testthat/test-methods.R
@@ -13,12 +13,10 @@ test_that("as data frame complete", {
   allf <- unlist(the_cff, recursive = TRUE, use.names = FALSE)
   expect_identical(length(allf), ncol(the_df))
 
-
   expect_identical(ncol(the_df), length(unique(names(the_df))))
 
   # As vector
   df_as_v <- as.character(as.vector(the_df[1, ]))
-
 
   expect_identical(allf, df_as_v)
   expect_snapshot(names(the_df))
@@ -38,8 +36,6 @@ test_that("as data frame partial", {
   allf <- unlist(subcff, recursive = TRUE, use.names = FALSE)
   expect_identical(length(allf), ncol(the_df))
   expect_identical(ncol(the_df), length(unique(names(the_df))))
-
-
 
   # To df references only
   subcff <- the_cff$references[[1]]
@@ -69,12 +65,10 @@ test_that("Convert a citation only", {
   allf <- unlist(the_cff, recursive = TRUE, use.names = FALSE)
   expect_identical(length(allf), ncol(the_df))
 
-
   expect_identical(ncol(the_df), length(unique(names(the_df))))
 
   # As vector
   df_as_v <- as.character(as.vector(the_df[1, ]))
-
 
   expect_identical(allf, df_as_v)
   expect_snapshot(names(the_df))
@@ -103,7 +97,6 @@ test_that("Convert authors only", {
   # As vector
   df_as_v <- as.character(as.vector(the_df[1, ]))
 
-
   expect_identical(allf, df_as_v)
   expect_snapshot(names(the_df))
 })
@@ -128,7 +121,6 @@ test_that("Convert list of authors", {
   # As vector
   df_as_v <- as.character(as.vector(the_df[1, ]))
 
-
   expect_identical(allf, df_as_v)
   expect_snapshot(names(the_df))
 })
@@ -136,7 +128,7 @@ test_that("Convert list of authors", {
 
 test_that("as.person method", {
   rvers <- getRversion()
-  skip_if(!grepl("^4.4", rvers), "Snapshot created with R 4.4.*")
+  skip_if(!grepl("^4.5", rvers), "Snapshot created with R 4.5.*")
   skip_on_cran()
   path <- system.file("examples/CITATION_complete.cff", package = "cffr")
 
@@ -230,18 +222,17 @@ test_that("head and tail", {
 test_that("toBibtex", {
   skip_on_cran()
 
-
   # Create several alternatives
-  descobj <- cff_read_description(system.file("examples/DESCRIPTION_basic",
+  descobj <- cff_read_description(system.file(
+    "examples/DESCRIPTION_basic",
     package = "cffr"
   ))
 
-  citobj <- cff_read_citation(system.file("examples/CITATION_basic",
+  citobj <- cff_read_citation(system.file(
+    "examples/CITATION_basic",
     package = "cffr"
   ))
-  newbib <- cff_read_bib(system.file("examples/example.bib",
-    package = "cffr"
-  ))
+  newbib <- cff_read_bib(system.file("examples/example.bib", package = "cffr"))
 
   full_cff <- merge_desc_cit(descobj, c(newbib, citobj))
   full_cff <- new_cff(full_cff)
@@ -261,7 +252,6 @@ test_that("toBibtex", {
   several <- toBibtex(full_cff$references)
   expect_s3_class(several, "Bibtex")
   expect_equal(sum(names(several) == "title"), 3)
-
 
   # One entry
   oneent <- toBibtex(full_cff$references[[1]])
@@ -284,7 +274,6 @@ test_that("toBibtex", {
   froml <- toBibtex(cff_read_bib_text(string))
   expect_equal(sum(names(froml) == "title"), 1)
 
-
   # Persons
 
   sev_auth <- as_cff_person(
@@ -292,8 +281,10 @@ test_that("toBibtex", {
   )
 
   comp <- c(
-    person("The Big Bopper"), person("Frank", "Sinatra"),
-    person("Dean", "Martin"), person("Sammy", "Davis, Jr.")
+    person("The Big Bopper"),
+    person("Frank", "Sinatra"),
+    person("Dean", "Martin"),
+    person("Sammy", "Davis, Jr.")
   )
 
   expect_length(sev_auth, 4)