+++ tests/testthat/test-utils-cff_create.R
@@ -1,5 +1,6 @@
 test_that("Parse Licenses on CRAN (Top 10)", {
-  cran_packs <- read.csv(system.file("extdata/cran_licenses.csv",
+  cran_packs <- read.csv(system.file(
+    "extdata/cran_licenses.csv",
     package = "cffr"
   ))
   cran_packs <- cran_packs[order(cran_packs$n, decreasing = TRUE), ]
@@ -8,11 +9,12 @@ test_that("Parse Licenses on CRAN (Top 10)", {
 
   # Tempfile DESCRIPTION that we use for modifications
   tmp <- tempfile("DESCRIPTION")
-  file.copy(system.file("examples/DESCRIPTION_basic", package = "cffr"),
+  file.copy(
+    system.file("examples/DESCRIPTION_basic", package = "cffr"),
     tmp,
     overwrite = TRUE
   )
-  for (i in seq_len(length(sel))) {
+  for (i in seq_along(sel)) {
     p <- desc::desc_set("License", sel[i], file = tmp)
     expect_silent(r <- cff_validate(cff_create(tmp), verbose = FALSE))
     expect_true(r)