+++ tests/testthat/test-utils-create.R
@@ -1,17 +1,16 @@
 test_that("Merge all DESCRIPTION files with CITATION_basic", {
   rvers <- getRversion()
-  skip_if(!grepl("^4.4", rvers), "Snapshot created with R 4.4.*")
+  skip_if(!grepl("^4.5", rvers), "Snapshot created with R 4.5.*")
   skip_on_cran()
 
   allfiles <- list.files(
     system.file("examples", package = "cffr"),
-    pattern = "^DESC", full.names = TRUE
+    pattern = "^DESC",
+    full.names = TRUE
   )
 
-  citpath <- system.file("examples/CITATION_basic",
-    package = "cffr"
-  )
-  for (i in seq_len(length(allfiles))) {
+  citpath <- system.file("examples/CITATION_basic", package = "cffr")
+  for (i in seq_along(allfiles)) {
     desc_cff <- cff_read_description(allfiles[i], gh_keywords = FALSE)
     generate_cit <- cff_safe_read_citation(allfiles[i], citpath)
     merged <- merge_desc_cit(desc_cff, generate_cit)
@@ -30,7 +29,7 @@ test_that("Check dependencies", {
   # Extract selected fields
   selected <- lapply(deps, function(x) {
     y <- x[names(x) %in% c("title", "url", "repository")]
-    return(y)
+    y
   })
 
   class(selected) <- "cff"